requirements.txt
PyQt5
keyboard
sounddevice
scipy
openai
pyperclip
pyautogui
python-dotenv
numpy
flask
supabase
